change profile to editprofile in bottom nav bar, user is not allowed to change first and lastname, user will enter edit profile fragment when clicking on the editprofile icn
