+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,6 @@
 
         <activity android:name=".CommentActivity"/>
         <activity android:name=".DetailPost" />
-        <activity android:name=".EditProfileActivity" />
         <activity android:name=".SignUpActivity" />
         <activity android:name=".MainActivity" />
         <activity android:name=".LoginActivity">
